BigDataApplication/requirements.txt
psycopg2-binary
matplotlib
pandas
numpy
seaborn
